Updated deprecated call of assertEquals in checkResultSupply method of TestEventRegistrationController.java

+++ Desktop/FoodTruckManagementSystem/src/ca/mcgill/ecse321/foodtruckmanagementsystem/controller/ItemController.java
@@ -126,7 +126,7 @@ public class ItemController {
 		PersistenceXStream.saveToXMLwithXStream(m);
 	}
 	
-	public void removeSupply(String name, int quantity) throws InvalidInputException{
+	public void removeSupply(String name, double quantity) throws InvalidInputException{
 		String error ="";
 		if ((name == null || name.trim().length() == 0))
 			error = error + "Supply name cannot be empty!";
